requirements.txt
fastapi==0.114.0
uvicorn[standard]==0.23.1
torch==2.1.0
torchvision==0.16.0
ftfy==6.1.1
regex==2025.10.3
tqdm==4.66.1
numpy==1.27.4
pillow==10.1.0
faiss-cpu==1.7.4
requests==2.32.1
